Remove unnecessary and inaccurate sentence from simple name resolution
The sentence attempts to simply state that a simple name from a
nested type that resolves to a name in an enclosing type must refer
to a static member - but it gets more complex than that when the
name refers to a method group.

Fortunately, this is all handled by effectively delegating to a
member access using the type name - that member access already
handles attempts to access instance members as if they were static.

Fixes #187