The ANTLR-isation of the lexical grammar. The C# "lexical" grammar both tokenises and parses the input and this requires using both lexical and parser ANTLR rules while producing a lexer which produces a token stream and not some AST of the input!
In this stage of ANTLR-isation the pre-processor is not tackled; the pre-processor definition is a context-free
grammar which is atypical for a lexical grammar and defining it with lexical rules does not produce the expected
result... That said the semantics of the pre-processor can be acheived in an ANTLR lexical grammar which includes
predicates/actions and that will be explored in a subsequent stage. The grammar in this stage can be used along
with a factored syntactical grammar to lex & parse C# source which does not include pre-processor constructs.

The majority of changes are to lexical-structure.md, many of the other files are only altered as the lexical
rule *Identifier* has been replaced by the parser rule *identifier*.