Fixed issue upsetting the automatic grammar extractor (I hope). Corrected Notes violating our style guidelines.
a) There are three rule groups; with primary rules *Input*, *Token* and *PP_Directive*; in the original "lexical"
grammar which together do not define lexical tokens but rather the lexical structure and pre-proccesing directives.
As is these rules are invalid as ANTLR lexical rules, and just switching them to be ANTLR parsing rules doesn't
work either.

In the first commit I had ANTLR-ised *Input* & *Token* groups, as the parser rule groups *input* and *token* but
had left the pre-processor rules as future work. However Bill's grammar extractor pointed out I'd violated an
ANTLR rule in doing so. I've made a small change to try to beat the grammar extractor, we'll see.

As before the pre-processor grammar remains invalid as ANTLR (and it was really always invalid as lexer rules)
but as a *description* it does its job as it always has.

(If people want the technical details ask on the ECMA list and I'll explain.)

b) Rex spotted I've violated our style guidelines for *Notes* by having some as "in paragraph". These have been
changed to be standalone indent paras.