This commit effectively disables the parts of the lexical grammar which are not acceptable to ANTLR
a) The `Input` group of rules which define the lexical structure rather than tokens

b) The `PP_Directive` group of rules which define the pre-processor directives. The rules as is
   define also of these as single tokens, which isn't what is intended!

c) The `Token` rule, and associated introduced `Any*` rules, which are used only by (a) and (b)

The disabling is accomplished by a liberal spreading of `fragment`.

In phase 2 the intention is to convert the pre-processor description to be compatible with ANTLR.