ANTLR-isation of parse/syntactic grammar
IMPORTANT: As PR#342 which ANTLR-ises the lexical & preprocessor grammars is not yet merged those
changes are NOT in the grammar in this PR. For the files changed by this PR this means you'll see
`Identifier` rather than `identifier` etc.

This PR makes the small number of changes to ANTLR-ise all but `primary_no_array_creation_expression`
which is being left in its mutual left recursive state for now/ever...

There is very little change to the descriptive text, these are mainly small grammar tweaks to remove
mutual left recursion.