Second pass at fixing #341
I *haven't* reinstated the reference within standard-library.md, as
it was anomalous there. (Our only other comment is for String.Format,
which refers to a later note, and for a much better reason than here.)

I've reinstated the reference from 9.6 to 11.7.3, but noted that
there aren't details.

I've included in 11.7.3 the fact that the rules for which
conversions will succeed are also implementation-dependent.