Initial draft for interpolated strings; the grammar is good, the prose a first draft.
Once the decision on how much of the ANLTR language we would use was made the grammar
was fairly straghtforward.

The prose describing the feature is more of a challenge as it is a language-level
surfacing of features of the Standard Library, and those features are either defined
in another Standard or not defined in any Standard but which we pretend are…

So the chanllenge is how much of the semantics should be included somewhere in this
Standard: what is a `FormattableString` vs. a `string` for the result type of an
*interpolated_string_expression*; what is the field width, what is the format, etc.?

This draft punts on some of these (e.g. the first one), includes details (e.g. for the the second
one), and even a new *informative* appendix, §C.4 (for the third one).

Is this balance right?

Well that depends on how we view our aim here…