Markdown lint fixes (#530)
* enable MD030

* fix MD030

These should all be whitespace-only changes.

* enable MD032

* skip 31, that will be a single PR. Use 32

* Fix MD032

Most of these are notes and examples. In most cases, the trailing *end note* was on a separate line not separated by a blank line. In many, they were at the end of the last bullet item. I normalized all to be on the last bullet item.

* Enable MD036

* re-disable MD036. Enable MD037

* Remove extra space in italicized words.

* enable MD038

* Fix MD038

These are inline code fences where one of the back-ticks was either missing, or misplaced.

* enable MD040

* Fix MD040

The only occurrences are in the tools/GetGrammar folder. These files are added by the tool, and insert the closing codefence. They are false positives. So, disable checking in that folder. We'll catch any mistakes there by linting the tool output on the automated PRs.

* enable MD041

* Update yml

Caught while testing.

* enable MD047

* enable MD050

* Fix MD050

Strong style should be consistent.