Allow output to be inferred from a section after the example instead of stated explicitly
This allows us to test that the claimed output in the standard is *really* the output. In various cases, this wasn't actually the case...

We still need ExpectedOutput at the moment for cases where the standard doesn't have the expected output in a triple-backtick-fenced block, but we could potentially aim to improve that over time.