Tweak undefined-behavior examples
In example PointerTypes1, the act of dereferencing `px1` and `px2` in `Main` is undefined. As such, the program might even abort! And even if the program executes, there is no point attempting to write out the underling `int` values, as they can be bogus, and certainly not predictable. (In my testing, *each* time I ran this program with output, one of the two values changed.)

So, I removed expectedOutput, changed the template, added some comments, and replaced the writes with dereferences to two new local variables.

For example PointerConversions1, I made no changes; I just activated this example.

For both examples, I added an HTML comment: `> <!-- Note: the behavior of this example is undefined. -->`.