one more (questionable) link fix
The text link says "unmanaged types", but the target is "pointer types". From both the context and the description in this paragraph, I think "unmanaged types" is a better target. However, this link has been to "pointer types" since C# 4.0 (or older). This one might be better left as is.