Clarify method group conversion
Fixes #552

First, remove the bullet regarding a compile time error if the method group is not compatible with the delegate type. That's now handled by applicable function member.

Earlier, point out that the conversion from a method group to a delegate includes the restrictions from the section on Applicable Function member that there must be an identity conversion from the method return type to the delegate's return type