Run quest nightly
In other repos, we've found that Quest can suffer from a race condition when a person updates multiple labels, triggering multiple runs in succession.

The best way to prevent that is to run nightly, during a time when it's unlikely that we're actively making updates. However, even if we are, this will be resilient because it will run once, and any updates made while it runs will update the following day.