Fix possible mistake in documentation-comments.md
In D.3.16 <summary> there is:
"This tag can be used to describe a type or a member of a type. Use <remarks> (§D.3.12) to describe the type itself."

And in D.3.12 <remarks> there is:
This tag is used to specify extra information about a type. Use <summary> (§D.3.16) to describe the type itself and the members of a type.

The description of the <summary> in both parts is OK. However in case of <remarks> the description in D.3.16 doesn't seem right when compared to the description in D.3.12.