Revision of §13.6.2 *Local variable declarations* to fix errors and tidy up.
This clause has evolved over time as first implicitly typed, and now in v7 ref locals, have been introduced; the result ending up
as something worthy of Heath Robinson or Professor Branestawm ;-) The result alas is also not entirely correct, the initial version
was (almost), but subsequent kinds of local variables don't follow exactly the same rules...

As a simple example of what ended up going wrong is the assertion:

> A local variable declaration that declares multiple variables is equivalent to multiple declarations of single variables with the same type and *ref_kind*.

and its associated example. This was (almost) correct originally, but is not correct for implicitly typed or ref local declarations where the
initialization cannot be split from the declaration. (It was “almost” is it didn't allow for *for_initializer* or *resource_acqusition* cases
where such a split is not possible either.)

The rework divides local variable declarations into three subgroups, there really are four but we can get away with three (whether
we should can be debated). These groups are directly represented in the reworked grammar, which also replaces a lot of conditional English
prose. The clause starts with the general rules and then has a section for each subgroup detailing just the rules that apply in that case.

The changes have been kept to the §13.6.2 clause; there are related issues in at least §7.7.1 Scopes:General and §13.14 The using statement – neither of these
for example define the scope of a variable introduced in a *resource_acqusition*. Such issues are left for future (pre-v8) work!