Add a general sub-clause (#957)
* Add a general sub-clause

In a recent PR, we broke apart local variable declarations into three different sections. When we did that, we created a clause that has both text and sub-clauses.

This PR fixes that.

I spoke offline with @RexJaeschke and he's making the appropriate edit to the Word doc for submission.

I'll cherry-pick this PR and the subsequent automated PR into `draft-v8` once this is merged.

* run section renumber locally

This will make it easier to cherry pick this once it's merged.