Create word artifact when a PR is merged to the main branch (#971)
* Generate the word doc as an artifact.

Add a workflow to generate the word doc and upload it as an artifact.

* updated deprecated commands.

`::set-output` is being deprecated. Replace it with `status=` syntax.

* Report status from converter run

* add an id

* YAML syntax

* reorder steps

* specify the shell

* path fixes

* remove unused array declaration

This isn't used by the tool, it reads all files in the folder, so just remove it.

Also, use the folder variable, rather than the hard-coded string.