Address async issues (#1027)
* "produces a result" vs. "return a value"

Fixes #857

For an async method, say "produces a result value" vs. "return a value" when describing the rules for the returned "task type".

* Add note about identity convertible

Fixes #878

I made this a note, rather than normative language. I read the existing text to allow the example in the issue (as does the roslyn implementation).

* Describe task cancelled state

Fixes #775

Add a bullet to the list specifying that an `OperationCanceledException` puts the returned task in the cancelled state.

This meant adding `OperationCanceledException` to the types in Annex C.

I chose this fix because all known task builders conform to this behavior. It seems reasonable to standardize it.

* Fixes #879

Clarify the required accessibility of any task builder type and its required members.

Note that we will need to update this text once dotnet/roslyn#71283 is fixed and merged.

* Add clarifying note.

The normative language on exactly matching accessibility is all that's strictly needed. But, I added a note about the implications for nested types and internal types.

* Fix missing exception type.

* Apply suggestions from code review

* Apply suggestions from code review

* Apply suggestions from code review

* Update standard/classes.md