Update dependencies, plus merge fix (#1134)
Two distinct changes here.

First, update all dependencies. Dependabot should have updated these, but that PR is failing the build. It appears that some updates weren't correctly applied.

Second, a quick work-around for the "update on merge" action so that it runs correctly. It's block, so I just copied the shell script and made mods. I'll take a bit longer to create a single script that reports errors and warnings correctly when the actions run not as part of a PR validation. I need to do some research there, and this is blocking us.