Preprocessor symbols for nullable reference types: Define a single nullable context. (#1123)
* Port text from #700

The first commit ports in the text from #700 substantially without change. The only editorial change is to move the clause on "Nullable directives" in lexical-structure before the clause on Pragma directives.

In addition, it fixes build warnings

* Remove link (which will need to get added)

Remove an Xref to a section that's not in this PR.

* Define a single nullable context

Alternative PR to #1108.

This branch builds on the work in #1108. This is an alternative approach that defines a single `nullable` context. The nullable context has two properties: `annotations` and `warnings` that can be enabled or disabled separately.

* Update per meeting comments in review

This commit addresses all online feedback.

* Respond to feedback from June meeting

I listened to the June meeting. I updated a bit of text, and added Jon's example  from the conversation.

* Update lexical-structure.md

* Apply suggestions from code review

* Update standard/lexical-structure.md