The compiler -> a compiler
Search for instances of "the compiler" and where the text describes what a compliant compiler is is required or allowed to do, replace "the compiler" with "a compiler".  Where necessary, include "shall" or "may", as appropriate.

There are still instances of "the compiler" in the standard. Some are in types.md. Those are addressed in #1125.  Others are where the text describes "the compiler generated" backing field, class, or method.