Unconstrained type parameters and `??` (#1323)
* Unconstrainted type parameters and `??`

Fixes #737

Add the language to the null-coalescing operator so that `??` can be used with an unconstrained type parameter.

* Revert "Unconstrainted type parameters and `??`"

This reverts commit a958a379933840153266ab31ce6e6e0780ace2f5.

* Do this the easy way.

I think this can be done the easy way. After reverting the original change, this commit removes the phrase "or a reference type" from the first bullet. By doing that, an unconstrained type parameter `T` is allowed. It's not a reference type.

I tested the case Nigel and Kalle described for using `??` with an unconstrained type parameter, and the existing rules produce the behavior they described.

I ran a few other cases to validate, and those worked correctly as well.

* Update standard/expressions.md

Co-authored-by: Nigel-Ecma <6654683+Nigel-Ecma@users.noreply.github.com>

* add exclusion on unmanaged types

Respond to Nigel's question

* respond to feedback

Change "a non-nullable value type" to "known to be a non-nullable value type".

Add an example to illustrate that `a ?? b` is `a` when the type of `a` is a non-nullable value type used as a type argument to an unconstrained type parameter.

* Update standard/expressions.md

Co-authored-by: Jon Skeet <skeet@pobox.com>

* extend the bulleted list.

* fix test harness issue

* Update standard/expressions.md

* Update standard/expressions.md

---------

Co-authored-by: Nigel-Ecma <6654683+Nigel-Ecma@users.noreply.github.com>
Co-authored-by: Jon Skeet <skeet@pobox.com>