Add language on cancellation
I stopped a bit short here. The more I dove into the cancellation protocol, the more it felt like this was an implementation spec, not a standard. Adding a good illustrative sample would mean showing the implementation of the async state machine for an async iterator. That seems to restricting.