Add support for using declarations and async using (#672)
* add support or using declarations and async using

* Add IAsyncDisposable to the library

* Update statements.md

* Add example annotation

* add annotation for examples

* Create SupportLocalVarDecl.cs

* fix section number

* fix links

* Update statements.md

* fix header label

* describe async using expansion

Include the rule that the result of `DisposeAsync` must be `await`ed.

* Update standard/statements.md

Co-authored-by: Nigel-Ecma <6654683+Nigel-Ecma@users.noreply.github.com>

* Address language in review comments

This commit address the comments in this review: https://github.com/dotnet/csharpstandard/pull/672#pullrequestreview-2908818087

except for any grammar changes.

Notes on where the edit differs (slightly) from the recommendation:

- The rule that a `using` statement can't declare an `out` variable applies to all forms. I added that in 13.14.1.
- I added a note that ref structs can't use the form of `await using`. It's a note because it follows from the normative restriction that `ref struct` types can't implement interfaces, but `await using` requires the resource type to implement `IAsyncDisposable`.
- The rules for `goto`, `break`, etc. are a note at the end of 13.14.1. They can be derived from the expanded forms of the `using` statement.
- The rule for a using declaration in a switch block remains. It's specific to a using declaration.

* Update grammar

Update grammar based on https://github.com/dotnet/csharpstandard/pull/672#pullrequestreview-2908818087

* refactor foreach into common sections

Refactor the common processing for `foreach` and `await foreach`. The common processing is the determination of the iteration type and such.

The lowering is still split into different sections. That could be combined, but as each is specific, I'm not sure it provides a simplification.

* Add attribute section for async iterator cancellation

In addition, add usings to the relevant template

* add normative text for async cancellation

* Address feedback in #606

Post meeting, a number of reviews were left on #606 to be addressed in this PR. All but one are addressed in this commit

* Clarify sync vs. async enumerable objects.

* Add missing note.

* address some feedback

Address some of the feedback items before the upcoming meeting.

* Apply suggestions from code review

Co-authored-by: Jon Skeet <skeet@pobox.com>

* Apply suggestions from code review

Co-authored-by: Joseph Musser <me@jnm2.com>
Co-authored-by: Jon Skeet <skeet@pobox.com>

* Interim checkin.

* Add language on cancellation

I stopped a bit short here. The more I dove into the cancellation protocol, the more it felt like this was an implementation spec, not a standard. Adding a good illustrative sample would mean showing the implementation of the async state machine for an async iterator. That seems to restricting.

* Remove `using` from duplicated grammar

* Apply suggestions from code review

Co-authored-by: Jon Skeet <skeet@pobox.com>

* Apply suggestions from code review

Co-authored-by: Jon Skeet <skeet@pobox.com>

* respond to feedback from July 30 meeting

* Apply suggestions from code review

Co-authored-by: Nigel-Ecma <6654683+Nigel-Ecma@users.noreply.github.com>

* Edits, part 1

Respond to review from @Nigel-Ecma on 8/15, *except* the summary comment.

* Address overall review comments

This might still be a bit hand-wavy, but better.

* Apply suggestions from code review

Co-authored-by: Joseph Musser <me@jnm2.com>

* respond to feedback.

* Apply suggestions from code review

Co-authored-by: Joseph Musser <me@jnm2.com>

* Add async dispose specification

* Update sample, fix build issue.

* Apply suggestions from code review

Co-authored-by: Jon Skeet <skeet@pobox.com>

* remaining review comments as of 9/2

* Update standard/attributes.md

Co-authored-by: Nigel-Ecma <6654683+Nigel-Ecma@users.noreply.github.com>

* Clarify dispose and finally

Clarify the semantics of DisposeAsync and Dispose for asynchronous enumerators.

* respond to final feedback.

* Apply suggestions from code review

Co-authored-by: Nigel-Ecma <6654683+Nigel-Ecma@users.noreply.github.com>

* Respond to feedback.

Update cancellation example. Make description of `finally` clauses consistent.

* Update sample and fix warnings

Update the sample per Nigel & Joseph's comments.

Also, update anchors that have changed due to the index and range merge.

* Update standard/attributes.md

Co-authored-by: Joseph Musser <me@jnm2.com>

* Update standard/attributes.md

* Update standard/attributes.md

* Update standard/attributes.md

* fix sample

* ensure sample compiles in C# 8

---------

Co-authored-by: Bill Wagner <wiwagn@microsoft.com>
Co-authored-by: Nigel-Ecma <6654683+Nigel-Ecma@users.noreply.github.com>
Co-authored-by: Jon Skeet <skeet@pobox.com>
Co-authored-by: Joseph Musser <me@jnm2.com>